feat: enrollment sql rendering [DHIS2-12323]
